+++ previous-versions/assess/RELEASES.md
@@ -401,3 +401,13 @@ Based on:
 - [go v0.27.0] previous-versions/assess
 ### Releases
 - [Go v0.27.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/assess/v0.27.0 - previous-versions/assess
+
+## 2023-10-01 00:15:02
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.92.2 (2.142.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.28.0] previous-versions/assess
+### Releases
+- [Go v0.28.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/assess/v0.28.0 - previous-versions/assess